+++ data/types/activity/physical/physical.go
@@ -3,8 +3,9 @@ package physical
 import (
 	"github.com/tidepool-org/platform/data"
 	"github.com/tidepool-org/platform/data/types"
-	dataTypesCommon "github.com/tidepool-org/platform/data/types/common"
+	"github.com/tidepool-org/platform/data/types/common"
 	"github.com/tidepool-org/platform/structure"
+	structureValidator "github.com/tidepool-org/platform/structure/validator"
 )
 
 const (
@@ -185,23 +186,26 @@ func ReportedIntensities() []string {
 type Physical struct {
 	types.Base `bson:",inline"`
 
-	ActivityType      *string                   `json:"activityType,omitempty" bson:"activityType,omitempty"`
-	ActivityTypeOther *string                   `json:"activityTypeOther,omitempty" bson:"activityTypeOther,omitempty"`
-	Aggregate         *bool                     `json:"aggregate,omitempty" bson:"aggregate,omitempty"`
-	Distance          *Distance                 `json:"distance,omitempty" bson:"distance,omitempty"`
-	Duration          *dataTypesCommon.Duration `json:"duration,omitempty" bson:"duration,omitempty"`
-	ElevationChange   *ElevationChange          `json:"elevationChange,omitempty" bson:"elevationChange,omitempty"`
-	Energy            *Energy                   `json:"energy,omitempty" bson:"energy,omitempty"`
-	Flight            *Flight                   `json:"flight,omitempty" bson:"flight,omitempty"`
-	Lap               *Lap                      `json:"lap,omitempty" bson:"lap,omitempty"`
-	Name              *string                   `json:"name,omitempty" bson:"name,omitempty"`
-	ReportedIntensity *string                   `json:"reportedIntensity,omitempty" bson:"reportedIntensity,omitempty"`
-	Step              *Step                     `json:"step,omitempty" bson:"step,omitempty"`
+	ActivityType      *string           `json:"activityType,omitempty" bson:"activityType,omitempty"`
+	ActivityTypeOther *string           `json:"activityTypeOther,omitempty" bson:"activityTypeOther,omitempty"`
+	Aggregate         *bool             `json:"aggregate,omitempty" bson:"aggregate,omitempty"`
+	Distance          *Distance         `json:"distance,omitempty" bson:"distance,omitempty"`
+	Duration          *common.Duration  `json:"duration,omitempty" bson:"duration,omitempty"`
+	ElevationChange   *ElevationChange  `json:"elevationChange,omitempty" bson:"elevationChange,omitempty"`
+	Energy            *Energy           `json:"energy,omitempty" bson:"energy,omitempty"`
+	Flight            *Flight           `json:"flight,omitempty" bson:"flight,omitempty"`
+	Lap               *Lap              `json:"lap,omitempty" bson:"lap,omitempty"`
+	Name              *string           `json:"name,omitempty" bson:"name,omitempty"`
+	ReportedIntensity *string           `json:"reportedIntensity,omitempty" bson:"reportedIntensity,omitempty"`
+	Step              *Step             `json:"step,omitempty" bson:"step,omitempty"`
+	EventID           *string           `json:"eventId,omitempty" bson:"eventId,omitempty"`
+	InputTime         *common.InputTime `bson:",inline"`
 }
 
 func New() *Physical {
 	return &Physical{
-		Base: types.New(Type),
+		Base:      types.New(Type),
+		InputTime: common.NewInputTime(),
 	}
 }
 
@@ -216,7 +220,7 @@ func (p *Physical) Parse(parser structure.ObjectParser) {
 	p.ActivityTypeOther = parser.String("activityTypeOther")
 	p.Aggregate = parser.Bool("aggregate")
 	p.Distance = ParseDistance(parser.WithReferenceObjectParser("distance"))
-	p.Duration = dataTypesCommon.ParseDuration(parser.WithReferenceObjectParser("duration"))
+	p.Duration = common.ParseDuration(parser.WithReferenceObjectParser("duration"))
 	p.ElevationChange = ParseElevationChange(parser.WithReferenceObjectParser("elevationChange"))
 	p.Energy = ParseEnergy(parser.WithReferenceObjectParser("energy"))
 	p.Flight = ParseFlight(parser.WithReferenceObjectParser("flight"))
@@ -224,6 +228,8 @@ func (p *Physical) Parse(parser structure.ObjectParser) {
 	p.Name = parser.String("name")
 	p.ReportedIntensity = parser.String("reportedIntensity")
 	p.Step = ParseStep(parser.WithReferenceObjectParser("step"))
+	p.EventID = parser.String("eventId")
+	p.InputTime.Parse(parser)
 }
 
 func (p *Physical) Validate(validator structure.Validator) {
@@ -266,6 +272,11 @@ func (p *Physical) Validate(validator structure.Validator) {
 	if p.Step != nil {
 		p.Step.Validate(validator.WithReference("step"))
 	}
+	validator.String("eventId", p.EventID)
+	if p.EventID != nil && p.Duration == nil {
+		validator.WithReference("duration").ReportError(structureValidator.ErrorValueNotExists())
+	}
+	p.InputTime.Validate(validator)
 }
 
 // IsValid returns true if there is no error in the validator
@@ -301,4 +312,5 @@ func (p *Physical) Normalize(normalizer data.Normalizer) {
 	if p.Step != nil {
 		p.Step.Normalize(normalizer.WithReference("step"))
 	}
+	p.InputTime.Normalize(normalizer.WithReference("inputTime"))
 }