+++ data/types/common/test/inputtime.go
@@ -3,24 +3,24 @@ package test
 import (
 	"time"
 
-	dataTypesCommon "github.com/tidepool-org/platform/data/types/common"
+	"github.com/tidepool-org/platform/data/types/common"
 	"github.com/tidepool-org/platform/test"
 
 	"github.com/tidepool-org/platform/pointer"
 )
 
-func NewInputTime() *dataTypesCommon.InputTime {
-	datum := dataTypesCommon.NewInputTime()
+func NewInputTime() *common.InputTime {
+	datum := common.NewInputTime()
 	timeReference := test.RandomTime()
 	datum.InputTime = pointer.FromString(timeReference.Format(time.RFC3339Nano))
 	return datum
 }
 
-func CloneInputTime(datum *dataTypesCommon.InputTime) *dataTypesCommon.InputTime {
+func CloneInputTime(datum *common.InputTime) *common.InputTime {
 	if datum == nil {
 		return nil
 	}
-	clone := dataTypesCommon.NewInputTime()
+	clone := common.NewInputTime()
 	clone.InputTime = pointer.CloneString(datum.InputTime)
 	return clone
 }